refactored animations to ToolboxWithObjects instead of Experience to improve load time, added a Suspense tag for the whole toolbox instead of each individual object
